Replace config manager with config controller
