--- app/scripts/lib/config-manager.js
@@ -1,62 +1,0 @@
-/* The config-manager is one of the oldest existing persistence objects in MetaMask.
- * Today its functionality has almost entirely been replaced by "Controllers".
- *
- * In a way, it is the last themeless controller, with just a few properties & functions.
- */
-
-module.exports = ConfigManager
-function ConfigManager (opts) {
-  // ConfigManager is observable and will emit updates
-  this._subs = []
-  this.store = opts.store
-}
-
-ConfigManager.prototype.setConfig = function (config) {
-  var data = this.getData()
-  data.config = config
-  this.setData(data)
-  this._emitUpdates(config)
-}
-
-ConfigManager.prototype.getConfig = function () {
-  var data = this.getData()
-  return data.config
-}
-
-ConfigManager.prototype.setData = function (data) {
-  this.store.putState(data)
-}
-
-ConfigManager.prototype.getData = function () {
-  return this.store.getState()
-}
-
-ConfigManager.prototype.setPasswordForgotten = function (passwordForgottenState) {
-  const data = this.getData()
-  data.forgottenPassword = passwordForgottenState
-  this.setData(data)
-}
-
-ConfigManager.prototype.setSeedWords = function (words) {
-  var data = this.getData()
-  data.seedWords = words
-  this.setData(data)
-}
-
-ConfigManager.prototype.subscribe = function (fn) {
-  this._subs.push(fn)
-  var unsubscribe = this.unsubscribe.bind(this, fn)
-  return unsubscribe
-}
-
-ConfigManager.prototype.unsubscribe = function (fn) {
-  var index = this._subs.indexOf(fn)
-  if (index !== -1) this._subs.splice(index, 1)
-}
-
-ConfigManager.prototype._emitUpdates = function (state) {
-  this._subs.forEach(function (handler) {
-    handler(state)
-  })
-}
-
